Move data into seperate file
