+++ updater/main.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 
 	"github.com/google/go-github/v28/github"
+	"github.com/icco/cron/sites"
 	"github.com/sirupsen/logrus"
 	"golang.org/x/oauth2"
 	apiv1 "k8s.io/api/core/v1"
@@ -20,152 +21,15 @@ type Config struct {
 	GithubToken string
 }
 
-type SiteMap struct {
-	Host       string
-	Owner      string
-	Repo       string
-	Deployment string
-}
-
 var (
 	c *Config
-
-	// AllSites contains a list of all domains I update from my code.
-	AllSites = []SiteMap{
-		{
-			Host:       "cacophony.natwelch.com",
-			Owner:      "icco",
-			Repo:       "cacophony",
-			Deployment: "cacophony",
-		},
-		{
-			Host:       "chartopia.app",
-			Owner:      "icco",
-			Repo:       "charts",
-			Deployment: "charts",
-		},
-		{
-			Host:       "code.natwelch.com",
-			Owner:      "icco",
-			Repo:       "code.natwelch.com",
-			Deployment: "code",
-		},
-		{
-			Host:       "cron.natwelch.com",
-			Owner:      "icco",
-			Repo:       "cron",
-			Deployment: "cron",
-		},
-		{
-			Host:       "etu.natwelch.com",
-			Owner:      "icco",
-			Repo:       "etu",
-			Deployment: "etu",
-		},
-		{
-			Host:       "goals.natwelch.com",
-			Owner:      "icco",
-			Repo:       "goals",
-			Deployment: "goals",
-		},
-		{
-			Host:       "gotak.app",
-			Owner:      "icco",
-			Repo:       "gotak",
-			Deployment: "gotak",
-		},
-		{
-			Host:       "graphql.natwelch.com",
-			Owner:      "icco",
-			Repo:       "graphql",
-			Deployment: "graphql",
-		},
-		{
-			Host:       "hello.natwelch.com",
-			Owner:      "icco",
-			Repo:       "hello",
-			Deployment: "hello",
-		},
-		{
-			Host:       "inspiration.natwelch.com",
-			Owner:      "icco",
-			Repo:       "inspiration",
-			Deployment: "inspiration",
-		},
-		{
-			Host:       "life.natwelch.com",
-			Owner:      "icco",
-			Repo:       "lifeline",
-			Deployment: "life",
-		},
-		{
-			Host:       "melandnat.com",
-			Owner:      "icco",
-			Repo:       "melandnat.com",
-			Deployment: "melandnat",
-		},
-		{
-			Host:       "natwelch.com",
-			Owner:      "icco",
-			Repo:       "natwelch.com",
-			Deployment: "natwelch",
-		},
-		{
-			Host:       "photos.natwelch.com",
-			Owner:      "icco",
-			Repo:       "photos",
-			Deployment: "photos",
-		},
-		{
-			Host:       "postmortems.app",
-			Owner:      "icco",
-			Repo:       "postmortems",
-			Deployment: "postmortems",
-		},
-		{
-			Host:       "quotes.natwelch.com",
-			Owner:      "icco",
-			Repo:       "crackquotes",
-			Deployment: "quotes",
-		},
-		{
-			Host:       "reportd.natwelch.com",
-			Owner:      "icco",
-			Repo:       "reportd",
-			Deployment: "reportd",
-		},
-		{
-			Host:       "resume.natwelch.com",
-			Owner:      "icco",
-			Repo:       "resume",
-			Deployment: "resume",
-		},
-		{
-			Host:       "sadnat.com",
-			Owner:      "icco",
-			Repo:       "sadnat.com",
-			Deployment: "sadnat",
-		},
-		{
-			Host:       "walls.natwelch.com",
-			Owner:      "icco",
-			Repo:       "wallpapers",
-			Deployment: "walls",
-		},
-		{
-			Host:       "writing.natwelch.com",
-			Owner:      "icco",
-			Repo:       "writing",
-			Deployment: "writing",
-		},
-	}
 )
 
 func UpdateWorkspaces(ctx context.Context, conf *Config) {
 	repoFmt := "gcr.io/icco-cloud/%s:%s"
 	c = conf
 
-	for _, r := range AllSites {
+	for _, r := range sites.All {
 		sha, err := c.GetSHA(ctx, r.Owner, r.Repo)
 		if _, ok := err.(*github.RateLimitError); ok {
 			c.Log.WithContext(ctx).WithError(err).Warn("hit rate limit")
